+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 - Client
   - Switch to 64 byte latency probes instead of 32 bytes
+  - Route liveness admin-down signalling and ignore stale remote-down messages
 
 - Device controller
   - Now accepts the config agent's version in the grpc GetConfig call and includes it as a label in the controller_grpc_getconfig_requests_total metric